Add DRONE_SERVER_HOST to server to fix commit status error
We were getting Drone error like

    error setting commit status

http://discourse.drone.io/t/drone-0-6-fails-to-set-ci-status-on-github/435
says this is due to not having this env variable on the server.

I think we want the ingress.hostname to be a secret, but if it's a
secret, I don't know how to get it into the ingress.yml as the host to
set up the ingress properly. So we want it as a value, but we don't
want it checked into values.yaml, and it can't be a secret, so ... we
specify it during start up? This is a pretty unsatisfying solution,
but eh, it works, so...